pyproject.toml
[project]
name = "simple-ai-agents"
version = "0.1.0"
description = "A minimal Azure AI Agent API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.2",
    "uvicorn[standard]>=0.29.0",
    "gunicorn>=23.0.0",
    "agent-framework-core>=1.0.0b251120",
    "agent-framework-azure-ai>=1.0.0b251120",
    "azure-identity>=1.19.0",
    "azure-ai-projects>=1.0.0b11",
    "azure-ai-inference>=1.0.0b7",
    "python-dotenv",
    "pydantic-settings>=2.0.0",
    "neo4j>=5.0.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    # Observability
    "azure-core-tracing-opentelemetry>=1.0.0b12",
    "azure-monitor-opentelemetry-exporter>=1.0.0b44",
    "azure-monitor-opentelemetry>=1.8.1",
]

[project.scripts]
start-agent = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neo4j_provider", "src/samples", "src/logging_config.py", "src/util.py", "src/agent.py", "src/neo4j_client.py", "src/vector_search.py", "src/main.py"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.uv.workspace]
members = [
    "data-pipeline",
]
